test: fix test-fs-utimes on non-Y2K38 file systems
Move Y2K38-specific parts of test-fs-utimes to test-fs-utimes-y2K38.js.
On non-Windows, check for Y2K38 support and skip if it is unsupported.

Fixes: https://github.com/nodejs/node/issues/36591

PR-URL: https://github.com/nodejs/node/pull/37707
